+++ lib/slam/icp/RefScanMakerLM.cpp
@@ -12,7 +12,7 @@ const Scan2D *RefScanMakerLM::MakeRefScan() {
 
   const auto &localMap = m_cloud_map_ptr->local_map();
 
-  for (unsigned i = 0; i < localMap.size(); ++i) {
+  for (size_t i = 0; i < localMap.size(); ++i) {
     const ScanPoint2D &point = localMap[i];
     refPoints.emplace_back(point);
   }
